+++ package.json
@@ -59,12 +59,12 @@
     ]
   },
   "devDependencies": {
-    "@storybook/addon-actions": "^6.3.0-alpha.35",
-    "@storybook/addon-essentials": "^6.3.0-alpha.35",
-    "@storybook/addon-links": "^6.3.0-alpha.35",
-    "@storybook/node-logger": "^6.3.0-alpha.35",
+    "@storybook/addon-actions": "^6.2.9",
+    "@storybook/addon-essentials": "^6.2.9",
+    "@storybook/addon-links": "^6.2.9",
+    "@storybook/node-logger": "^6.2.9",
     "@storybook/preset-create-react-app": "^3.1.7",
-    "@storybook/react": "^6.3.0-alpha.35",
+    "@storybook/react": "^6.2.9",
     "@typescript-eslint/eslint-plugin": "^4.24.0",
     "@typescript-eslint/parser": "^4.24.0",
     "babel-loader": "8.1.0",
