+++ src/components/molecules/SectionAddForm/SectionAddForm.stories.tsx
@@ -17,9 +17,4 @@ Basic.args = {
     { id: 1, name: '도비1', upStationId: 1, downStationId: 2, distance: 10 },
     { id: 1, name: '도비1', upStationId: 1, downStationId: 2, distance: 10 },
   ],
-  stationList: [
-    { id: 1, name: '도비1' },
-    { id: 2, name: '도비2' },
-    { id: 3, name: '도비3' },
-  ],
 };
